+++ lib/main.dart
@@ -4,6 +4,7 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:mobile/firebase_options.dart';
 import 'package:mobile/screen/home_page.dart';
 import 'package:mobile/screen/profile.dart';
+import 'package:mobile/screen/profile_edit.dart';
 // import 'package:firebase_core/firebase_core.dart';
 
 import 'screen/sigup.dart';
