+++ lib/screen/profile.dart
@@ -1,100 +1,45 @@
+import 'package:firebase_auth/firebase_auth.dart';
+import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter/widgets.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
-import 'package:mobile/screen/profile2.dart';
+import 'package:mobile/screen/login.dart';
 
-class ProfileScreen extends HookWidget {
+class ProfileScreen extends StatefulWidget {
+  const ProfileScreen({super.key});
+
+  @override
+  State<ProfileScreen> createState() => _ProfileScreenState();
+}
+
+class _ProfileScreenState extends State<ProfileScreen> {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
-        title: Text('Profile 編集'),
+        title: Text('Profile'),
       ),
-      body: Column(
-        children: [
-          Row(
-            mainAxisAlignment: MainAxisAlignment.end,
-            children: [
-              Padding(
-                padding: EdgeInsets.only(right: 20),
-                child: ElevatedButton(
-                  child: Text("編集を終了"),
-                  onPressed: () {
-                    Navigator.push(
-                      context,
-                      MaterialPageRoute(builder: (context) => Profile2Screen()),
-                    );
-                  },
-                ),
-              ),
-            ],
-          ),
-          Padding(
-            padding: EdgeInsets.only(top: 50),
-            child: Center(
-              child: Icon(Icons.account_circle, size: 96), // アイコンを配置
+      body: Center(
+        child: Column(
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: [
+            ElevatedButton(
+              onPressed: () async {
+                await FirebaseAuth.instance
+                    .signOut()
+                    .then((_) => Navigator.pushNamed(context, "/logIn"));
+              },
+              child: Text('Log Out'),
             ),
-          ),
-          Padding(
-            padding:
-                EdgeInsets.only(right: 150, left: 150, bottom: 20), //左、右、上、下
-            child: TextField(
-              textAlign: TextAlign.center, // テキストを中央揃えにする
-              style: TextStyle(fontSize: 24),
-              decoration: InputDecoration(
-                hintText: "name",
-              ),
-            ),
-          ),
-          Padding(
-            padding: EdgeInsets.symmetric(
-                horizontal: 50, vertical: 10), //horizonta:左右、vertical上下
-            child: TextField(
-              decoration: InputDecoration(
-                icon: Icon(Icons.person),
-                hintText: "nickname",
-              ),
-            ),
-          ),
-          Padding(
-            padding: EdgeInsets.symmetric(
-                horizontal: 50, vertical: 10), //horizonta:左右、vertical上下
-            child: TextField(
-              decoration: InputDecoration(
-                icon: Icon(Icons.cake),
-                hintText: "birthday",
-              ),
-            ),
-          ),
-          Padding(
-            padding: EdgeInsets.symmetric(
-                horizontal: 50, vertical: 10), //horizonta:左右、vertical上下
-            child: TextField(
-              decoration: InputDecoration(
-                icon: Icon(Icons.edit),
-                hintText: "comment",
-              ),
-            ),
-          ),
-          Padding(
-            padding: EdgeInsets.symmetric(
-                horizontal: 50, vertical: 10), //horizonta:左右、vertical上下
-            child: TextField(
-              decoration: InputDecoration(
-                icon: Icon(Icons.directions_walk),
-                hintText: "have visited",
-              ),
-            ),
-          ),
-          Padding(
-            padding: EdgeInsets.symmetric(
-                horizontal: 50, vertical: 20), //horizonta:左右、vertical上下
-            child: Text(
-              "書いた記事",
-              style: TextStyle(fontSize: 25),
-            ),
-          ),
-        ],
+            ElevatedButton(
+              onPressed: () async {
+                await FirebaseAuth.instance
+                    .signOut()
+                    .then((_) => Navigator.pushNamed(context, "/profile"));
+              },
+              child: Text('Profile'),
+            )
+          ],
+        ),
       ),
     );
   }
